+++ CHANGELOG.md
@@ -2,12 +2,14 @@
 此專案內所有值得被紀錄的更動都應該被紀錄於此文件。
 
 格式基於 [Keep a Changelog](http://keepachangelog.com/zh-TW/1.0.0/)
-這個項目堅持 [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
+這個項目堅持 [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
 ## 0.0.2 - 2018-08-10
 ### Added
 - 新增 helper 資料夾
 - 在 helper 下新增 logger 模組及新增 logger 資料夾
+### Changed
+- 將 logger 資料夾改成 log
 
 ## 0.0.1 - 2018-08-05
 ### Added